Separate verifies and filter locals, modified all db managers
