+++ M2/Back-end/pet_shop/src/db/sell_module/db_cart_item_manager.py
@@ -1,5 +1,6 @@
 from sqlalchemy import insert, select, delete, update, and_
-from src.db.utils_db.helpers import _filter_locals, _verify_user_own_cart, _verify_amount_product
+from src.db.utils_db.helpers import _filter_locals
+from src.db.utils_db.verifies import _verify_user_own_cart, _verify_amount_product
 from src.utils.api_exception import APIException
