+++ M2/Back-end/pet_shop/src/db/user_module/db_address_manager.py
@@ -1,6 +1,8 @@
 from sqlalchemy import insert, select, delete, update, and_
-from src.db.utils_db.helpers import _filter_locals, _verify_user_own_address
+from src.db.utils_db.helpers import _filter_locals
+from src.db.utils_db.verifies import _verify_user_own_address
 from src.utils.api_exception import APIException
+
 
 
 class DbAddressManager:
